Initial commit for SQL injection with JPA, Hibernate and Spring Data at one place
