+++ persistence-modules/sqlinjection/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+spring.datasource.url=jdbc:h2:mem:test
+spring.jpa.show-sql=false
+logging.level.com=error
