+++ persistence-modules/sqlinjection/src/main/resources/employee.cfg.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE hibernate-configuration PUBLIC
+        "-//Hibernate/Hibernate Configuration DTD 3.0//EN"
+        "http://www.hibernate.org/dtd/hibernate-configuration-3.0.dtd">
+
+
+<hibernate-configuration>
+
+	<session-factory>
+		<property name="connection.driver_class">org.h2.Driver</property>
+		<property name="connection.url">jdbc:h2:mem:test</property>
+		<property name="connection.username">sa</property>
+		<property name="hibernate.default_schema">PUBLIC</property>
+		<property name="dialect">org.hibernate.dialect.H2Dialect</property>
+		<property name="show_sql">false</property>
+		<property name="hbm2ddl.auto">create</property>
+
+		<mapping class="com.baeldung.sqlinjection.model.Employee" />
+
+	</session-factory>
+</hibernate-configuration>
