--- persistence-modules/sqlinjection2/.gitignore
@@ -1,1 +1,0 @@
-/bin/
