--- persistence-modules/sqlinjection2/src/main/java/com/baeldung/model/Employee.java
@@ -1,52 +1,0 @@
-package com.baeldung.model;
-
-import javax.persistence.*;
-
-
-@Entity
-@Table(name = "tbl_employee")
-public class Employee {
-        @Id
-        @GeneratedValue(strategy = GenerationType.IDENTITY)
-        private Long id;
-        private String name;
-
-        public String getName() {
-                return name;
-        }
-
-        public void setName(String name) {
-                this.name = name;
-        }
-
-        @Override
-        public String toString() {
-                return "Employee [id=" + id + ", name=" + name + "]";
-        }
-
-        @Override
-        public int hashCode() {
-                final int prime = 31;
-                int result = 1;
-                result = prime * result + ((id == null) ? 0 : id.hashCode());
-                return result;
-        }
-
-        @Override
-        public boolean equals(final Object obj) {
-                if (this == obj)
-                        return true;
-                if (obj == null)
-                        return false;
-                if (getClass() != obj.getClass())
-                        return false;
-                final Employee other = (Employee) obj;
-                if (id == null) {
-                        if (other.id != null)
-                                return false;
-                } else if (!id.equals(other.id))
-                        return false;
-                return true;
-        }
-
-}
