--- persistence-modules/sqlinjection2/src/main/resources/application.properties
@@ -1,7 +1,0 @@
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://127.0.0.1:3306/baeldung?autoReconnect=true&useSSL=false
-spring.datasource.username=baeldung
-spring.datasource.password=YourPassword
-spring.jpa.show-sql=true
-logging.level.=ERROR
-hibernate.dialect=org.hibernate.dialect.MySQLDialect
